+++ Eter-MC++/Eter-MC++/Coordinates.h
@@ -3,16 +3,16 @@ class Coordinates
 {
 public:
 	Coordinates();
-	Coordinates(int x, int y);
+	Coordinates(short x, short y);
 	Coordinates(const Coordinates& other);
 	size_t operator() (const Coordinates& key) const;
 	bool operator== (const Coordinates& other) const;
-	int GetX() const;
-	int GetY() const;
-	void SetX(int x);
-	void SetY(int y);
+	short GetX() const;
+	short GetY() const;
+	void SetX(short x);
+	void SetY(short y);
 
 private:
-	int m_x, m_y;
+	short m_x, m_y;
 };
